prevent job from being deleted accidentally;

+++ pages/job.vue
@@ -411,10 +411,13 @@ export default {
 			this.toggleEditor();
 		},
 		deleteJob() {
-			this.$store.dispatch('job/deleteJob', this.job.id);
-			this.$router.push('/');
+			if (confirm('Удалить '+this.job.id+'?')) {
+				this.$store.dispatch('job/deleteJob', this.job.id);
+				this.$router.push('/');
+			}
 		},
 		toggleEditor: function() {
+			console.log('gggg')
 			clearInterval(this.watcher);
 			this.editorVisibility = !this.editorVisibility;
 		},
